Crate binary_space_partition

Binary Space Partitioning (BSP)

Provides an abstract BspNode structure, which can be seen as a tree. Useful for quickly ordering polygons along a particular view vector. Is not tied to a particular math library.

Structs

BspNode
A node in the BspTree, which can be considered a tree itself.

Enums

PlaneCut
The result of one plane being cut by another one. The "cut" here is an attempt to classify a plane as being in front or in the back of another one.

Traits

Plane
A plane abstracted to the matter of partitioning.

Functions

add_side
Add a list of planes to a particular front/end branch of some root node.