Enum binary_space_partition::PlaneCut

pub enum PlaneCut<T> {
    Sibling(T),
    Cut {
        front: Vec<T>,
        back: Vec<T>,
    },
}

The result of one plane being cut by another one. The "cut" here is an attempt to classify a plane as being in front or in the back of another one.

Variants
Sibling(T)

The planes are one the same geometrical plane.

Cut

Planes are different, thus we can either determine that our plane is completely in front/back of another one, or split it into these sub-groups.

Fields of Cut

front: Vec<T>

Sub-planes in front of the base plane.

back: Vec<T>

Sub-planes in the back of the base plane.