Trait binary_space_partition::Plane

pub trait Plane: Sized + Clone {
    fn cut(&self, _: Self) -> PlaneCut<Self>;

    fn is_aligned(&self, _: &Self) -> bool;
}

A plane abstracted to the matter of partitioning.

Required methods

fn cut(&self, _: Self) -> PlaneCut<Self>

Try to cut a different plane by this one.

fn is_aligned(&self, _: &Self) -> bool

Check if a different plane is aligned in the same direction as this one.

Loading content...